+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/utils/søknadsdataToApiData/getApiDataFromSøknadsdata.ts
@@ -1,6 +1,5 @@
-import { getAttachmentURLBackend } from '@navikt/sif-common';
 import { Attachment } from '@navikt/sif-common-core-ds/src/types';
-import { attachmentIsUploadedAndIsValidFileFormat } from '@navikt/sif-common-core-ds/src/utils/attachmentUtils';
+import { getAttachmentsApiData } from '@navikt/sif-common-core-ds/src/utils/attachmentUtils';
 import { SøknadApiData } from '../../types/søknadApiData/SøknadApiData';
 import { SituasjonSøknadsdata, Søknadsdata } from '../../types/søknadsdata/Søknadsdata';
 import { getArbeidsgivereApiDataFromSøknadsdata } from './getArbeidsgivereApiDataFromSøknadsdata';
@@ -8,15 +7,6 @@ import { getDataBruktTilUtledning } from './getDataBruktTilUtledning';
 import { getDineBarnApiDataFromSøknadsdata } from './getDineBarnApiDataFromSøknadsdata';
 import { getMedlemskapApiDataFromSøknadsdata } from './getMedlemskapApiDataFromSøknadsdata';
 import { getUtenlansoppholdApiDataFromSøknadsdata } from './getUtenlandsoppholdApiDataFromSøknadsdata';
-
-const getVedleggApiData = (vedlegg?: Attachment[]): string[] => {
-    if (!vedlegg || vedlegg.length === 0) {
-        return [];
-    }
-    return vedlegg
-        .filter(attachmentIsUploadedAndIsValidFileFormat)
-        .map(({ url }) => (url ? getAttachmentURLBackend(url) : ''));
-};
 
 const getArbeidsforholdDokumenter = (situasjon: SituasjonSøknadsdata): Attachment[] => {
     const dokumenter: Attachment[] = [];
@@ -51,9 +41,9 @@ export const getApiDataFromSøknadsdata = (
         opphold: getUtenlansoppholdApiDataFromSøknadsdata(språk, fravær),
         bosteder: getMedlemskapApiDataFromSøknadsdata(språk, medlemskap),
         vedlegg: [
-            ...getVedleggApiData(deltBosted?.vedlegg),
-            ...getVedleggApiData(legeerklæring?.vedlegg),
-            ...getVedleggApiData(getArbeidsforholdDokumenter(situasjon)),
+            ...getAttachmentsApiData(deltBosted?.vedlegg),
+            ...getAttachmentsApiData(legeerklæring?.vedlegg),
+            ...getAttachmentsApiData(getArbeidsforholdDokumenter(situasjon)),
         ],
         dataBruktTilUtledningAnnetData: JSON.stringify(getDataBruktTilUtledning(søknadsdata)),
     };