+++ apps/omsorgspengesoknad/dev-server.cjs
@@ -39,11 +39,7 @@ async function injectDecorator(filePath) {
 }
 
 const startServer = async () => {
-    server.get('/health/isAlive', (req, res) => res.sendStatus(200));
-    server.get('/health/isReady', (req, res) => res.sendStatus(200));
-
     const indexHtmlPath = path.resolve(__dirname, 'index.html');
-
     const htmlWithDecoratorInjected = await injectDecorator(indexHtmlPath);
 
     const renderedHtml = htmlWithDecoratorInjected.replaceAll(