+++ apps/omsorgspengesoknad/e2e/playwright/tests/fylle-ut-søknad-med-annet-barn.spec.ts
@@ -37,7 +37,8 @@ test('Fyll ut søknad med annet barn', async ({ page }) => {
     /** Legeerklæring */
     await page.getByRole('heading', { level: 1, name: 'Legeerklæring' });
     await page.locator('input[name="vedlegg"]').setInputFiles('./e2e/playwright/files/navlogopng.png');
-    await expect(await page.getByText('Fjern').count()).toEqual(1);
+    const listItems = await page.getByText('navlogopng.png(2.31 KB)');
+    await expect(listItems).toHaveCount(1);
     await page.getByRole('button', { name: 'Neste steg', exact: true }).click();
 
     /** Oppsummering */