+++ apps/pleiepenger-i-livets-sluttfase-soknad/src/app/søknad/steps/oppsummering/components/PleietrengendePersonSummary.tsx
@@ -1,11 +1,11 @@
 import { FormSummary } from '@navikt/ds-react';
-import AttachmentList from '@navikt/sif-common-core-ds/src/components/attachment-list/AttachmentList';
 import { Attachment } from '@navikt/sif-common-core-ds/src/types/Attachment';
 import EditStepLink from '@navikt/sif-common-soknad-ds/src/components/edit-step-link/EditStepLink';
 import { FødselsnummerSvar } from '@navikt/sif-common-ui';
 import { ISODateToDate, prettifyDate } from '@navikt/sif-common-utils';
 import { AppText } from '../../../../i18n';
 import { FlereSokereApiData, PleietrengendeApi } from '../../../../types/søknadApiData/SøknadApiData';
+import AttachmentList from '@navikt/sif-common-core-ds/src/components/attachment-list/AttachmentList';
 
 interface Props {
     pleietrengende: PleietrengendeApi;