+++ apps/pleiepenger-sykt-barn/src/app/validation/fieldValidations.ts
@@ -1,9 +1,3 @@
-import { Attachment } from '@navikt/sif-common-core-ds/src/types/Attachment';
-import {
-    attachmentHasBeenUploaded,
-    getTotalSizeOfAttachments,
-    MAX_TOTAL_ATTACHMENT_SIZE_BYTES,
-} from '@navikt/sif-common-core-ds/src/utils/attachmentUtils';
 import { YesOrNo } from '@navikt/sif-common-formik-ds/src';
 import datepickerUtils from '@navikt/sif-common-formik-ds/src/components/formik-datepicker/datepickerUtils';
 import {
@@ -39,10 +33,6 @@ dayjs.extend(minMax);
 dayjs.extend(isoWeek);
 
 export enum AppFieldValidationErrors {
-    'samlet_storrelse_for_hoy' = 'validation.samlet_storrelse_for_hoy',
-    'legeerklæring_mangler' = 'legeerklæring.mangler',
-    'legeerklæring_forMangeFiler' = 'legeerklæring.forMangeFiler',
-
     'arbeidsforhold_timerUgyldig_under_1_prosent' = 'timerUgyldig_under_1_prosent',
     'arbeidsforhold_timerUgyldig_under_0_prosent' = 'timerUgyldig_under_0_prosent',
     'arbeidsforhold_timerUgyldig_over_99_prosent' = 'timerUgyldig_over_99_prosent',
@@ -169,18 +159,6 @@ export const validateFerieuttakIPerioden = (
     }
     if (dateRangesExceedsRange(dateRanges, periode)) {
         return AppFieldValidationErrors.ferieuttak_utenfor_periode;
-    }
-    return undefined;
-};
-
-export const validateLegeerklæring = (attachments: Attachment[]): ValidationResult<ValidationError> => {
-    const uploadedAttachments = attachments.filter((attachment) => attachmentHasBeenUploaded(attachment));
-    const totalSizeInBytes: number = getTotalSizeOfAttachments(attachments);
-    if (totalSizeInBytes > MAX_TOTAL_ATTACHMENT_SIZE_BYTES) {
-        return AppFieldValidationErrors.samlet_storrelse_for_hoy;
-    }
-    if (uploadedAttachments.length > 100) {
-        return AppFieldValidationErrors.legeerklæring_forMangeFiler;
     }
     return undefined;
 };