+++ packages/sif-common-core-ds/src/components/formik-attachment-form/parts/AttachmentTotalSizeAlert.tsx
@@ -1,19 +1,23 @@
 import { Alert, Link } from '@navikt/ds-react';
 import { CoreText } from '../../../i18n/common.messages';
 
-const AttachmentTotalSizeAlert = ({ uploadLaterURL }: { uploadLaterURL: string }) => {
+const AttachmentTotalSizeAlert = ({ uploadLaterURL }: { uploadLaterURL?: string }) => {
     return (
         <Alert variant="warning">
-            <CoreText
-                id="@core.formik-attachment-form.alert.totalSize"
-                values={{
-                    Lenke: (children: React.ReactNode) => (
-                        <Link target={'_blank'} rel={'noopener noreferrer'} href={uploadLaterURL}>
-                            {children}
-                        </Link>
-                    ),
-                }}
-            />
+            {uploadLaterURL ? (
+                <CoreText
+                    id="@core.formik-attachment-form.alert.totalSize"
+                    values={{
+                        Lenke: (children: React.ReactNode) => (
+                            <Link target={'_blank'} rel={'noopener noreferrer'} href={uploadLaterURL}>
+                                {children}
+                            </Link>
+                        ),
+                    }}
+                />
+            ) : (
+                <CoreText id="@core.formik-attachment-form.alert.totalSize.noLink" />
+            )}
         </Alert>
     );
 };