+++ apps/endringsmelding-pleiepenger/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @navikt/endringsmelding-pleiepenger
 
+## 2.32.6
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+-   Updated dependencies
+    -   @navikt/sif-common-amplitude@2.26.19
+    -   @navikt/sif-common-formik-ds@1.12.25
+    -   @navikt/sif-common-soknad-ds@14.0.20
+    -   @navikt/sif-common-forms-ds@10.1.19
+    -   @navikt/sif-common-core-ds@9.18.26
+    -   @navikt/sif-common-sentry@0.27.17
+    -   @navikt/sif-common-utils@3.47.20
+
+## 2.32.5
+
+### Patch Changes
+
+-   Legge til K9_BRUKERDIALOG_PROSESSERING_API_URL i appSettings
+
 ## 2.32.4
 
 ### Patch Changes